django/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
cachetools==4.2.4
decorator==4.4.2
Django==4.1.6
django-cors-headers==3.14.0
django-unicorn==0.49.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
orjson==3.8.5
PyJWT==2.6.0
pytz==2022.7.1
shortuuid==1.0.11
soupsieve==2.3.2.post1
sqlparse==0.4.3
